pyproject.toml
[project]
name = "portfolio-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-tagulous>=2.1.0",
    "django>=5.1.4",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "ruff>=0.8.2",
]
